+++ packages/tokens/src/rwa/mod.rs
@@ -1,6 +1,7 @@
 //! TODO: describe the module
 
-pub mod modules;
+pub mod claim_topics_and_issuers;
+pub mod compliance;
 
 /// TODO: describe the trait
 pub trait RWA {}
