--- packages/tokens/src/rwa/modules/mod.rs
@@ -1,2 +1,0 @@
-pub mod claim_topics_and_issuers;
-pub mod compliance;
